requirements.txt
Django~=5.1.5
drf-yasg~=1.21.9
djangorestframework~=3.15.2
celery~=5.4.0
python-dotenv~=1.0.1
django-celery-beat~=2.7.0
django-filter~=25.1
selenium~=4.29.0
bs4~=0.0.2
beautifulsoup4~=4.13.3
webdriver-manager~=4.0.2
Scrapy~=2.12.0
itemadapter~=0.11.0
mysqlclient
